Resolve some comments
